Source hash: cac84b5e057a074da6a72e2a8bbcc70ec0a09668
Remaining commits: 0

+++ chia/consensus/block_rewards.py
@@ -17,17 +17,17 @@ def calculate_pool_reward(height: uint32) -> uint64:
     """
 
     if height == 0:
-        return uint64(int((7 / 8) * 21000000 * _mojo_per_chia))
+        return uint64((7 / 8) * 21000000 * _mojo_per_chia)
     elif height < 3 * _blocks_per_year:
-        return uint64(int((7 / 8) * 2 * _mojo_per_chia))
+        return uint64((7 / 8) * 2 * _mojo_per_chia)
     elif height < 6 * _blocks_per_year:
-        return uint64(int((7 / 8) * 1 * _mojo_per_chia))
+        return uint64((7 / 8) * 1 * _mojo_per_chia)
     elif height < 9 * _blocks_per_year:
-        return uint64(int((7 / 8) * 0.5 * _mojo_per_chia))
+        return uint64((7 / 8) * 0.5 * _mojo_per_chia)
     elif height < 12 * _blocks_per_year:
-        return uint64(int((7 / 8) * 0.25 * _mojo_per_chia))
+        return uint64((7 / 8) * 0.25 * _mojo_per_chia)
     else:
-        return uint64(int((7 / 8) * 0.125 * _mojo_per_chia))
+        return uint64((7 / 8) * 0.125 * _mojo_per_chia)
 
 
 def calculate_base_farmer_reward(height: uint32) -> uint64:
@@ -40,14 +40,14 @@ def calculate_base_farmer_reward(height: uint32) -> uint64:
     rates increase continuously.
     """
     if height == 0:
-        return uint64(int((1 / 8) * 21000000 * _mojo_per_chia))
+        return uint64((1 / 8) * 21000000 * _mojo_per_chia)
     elif height < 3 * _blocks_per_year:
-        return uint64(int((1 / 8) * 2 * _mojo_per_chia))
+        return uint64((1 / 8) * 2 * _mojo_per_chia)
     elif height < 6 * _blocks_per_year:
-        return uint64(int((1 / 8) * 1 * _mojo_per_chia))
+        return uint64((1 / 8) * 1 * _mojo_per_chia)
     elif height < 9 * _blocks_per_year:
-        return uint64(int((1 / 8) * 0.5 * _mojo_per_chia))
+        return uint64((1 / 8) * 0.5 * _mojo_per_chia)
     elif height < 12 * _blocks_per_year:
-        return uint64(int((1 / 8) * 0.25 * _mojo_per_chia))
+        return uint64((1 / 8) * 0.25 * _mojo_per_chia)
     else:
-        return uint64(int((1 / 8) * 0.125 * _mojo_per_chia))
+        return uint64((1 / 8) * 0.125 * _mojo_per_chia)
